pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-summarization"
version = "0.1.0"
description = "Multi-modal Summarization: Combining text summarization with image captioning"
authors = [
    {name = "AI Projects", email = "projects@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "tensorboard>=2.13.0",
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.20.0",
    "open_clip_torch>=2.20.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "Pillow>=10.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "sacrebleu>=2.3.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "nltk>=3.8.0",
    "faiss-cpu>=1.7.4",
    "hnswlib>=0.7.0",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
    "wandb>=0.15.0",
    "mlflow>=2.5.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "S104", "S108", "S110", "S112", "S311", "S603", "S607"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
